scripts/files_for_summaries/requirements.txt
functions-framework==3.8.2
google-cloud-storage==3.1.0